[#62][hub] refactor: dto 위치 이동
