+++ hub/src/main/java/com/faster/hub/app/hub/application/dto/request/DeleteHubApplicationRequestDto.java
@@ -1,4 +1,4 @@
-package com.faster.hub.app.hub.application.dto;
+package com.faster.hub.app.hub.application.dto.request;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
